+++ app/src/ask-pass/ask-pass.ts
@@ -1,5 +1,5 @@
 import { getKeyForEndpoint } from '../lib/auth'
-import { TokenStore } from '../lib/dispatcher/token-store'
+import { TokenStore } from '../lib/stores/token-store'
 
 /** Parse the GIT_ASKPASS prompt and determine the appropriate response. */
 export async function responseForPrompt(
