+++ app/src/lib/menu-update.ts
@@ -2,7 +2,7 @@ import { MenuIDs } from '../main-process/menu'
 import { merge } from './merge'
 import { IAppState, SelectionType } from '../lib/app-state'
 import { Repository } from '../models/repository'
-import { CloningRepository } from './dispatcher'
+import { CloningRepository } from './stores/cloning-repositories-store'
 import { TipState } from '../models/tip'
 import { updateMenuState as ipcUpdateMenuState } from '../ui/main-process-proxy'
 import { AppMenu, MenuItem } from '../models/app-menu'
