+++ app/src/ui/history/commit-list.tsx
@@ -3,7 +3,7 @@ import { Repository } from '../../models/repository'
 import { Commit } from '../../models/commit'
 import { CommitListItem } from './commit-list-item'
 import { List } from '../list'
-import { IGitHubUser } from '../../lib/dispatcher'
+import { IGitHubUser } from '../../lib/databases'
 
 const RowHeight = 48
