+++ app/src/ui/history/commit-summary.tsx
@@ -4,7 +4,7 @@ import * as classNames from 'classnames'
 import { FileChange } from '../../models/status'
 import { Octicon, OcticonSymbol } from '../octicons'
 import { RichText } from '../lib/rich-text'
-import { IGitHubUser } from '../../lib/dispatcher'
+import { IGitHubUser } from '../../lib/databases'
 import { Repository } from '../../models/repository'
 import { Avatar } from '../lib/avatar'
 import { Commit } from '../../models/commit'
