+++ app/src/ui/history/sidebar.tsx
@@ -2,7 +2,8 @@ import * as React from 'react'
 import { CommitList } from './commit-list'
 import { Repository } from '../../models/repository'
 import { Commit } from '../../models/commit'
-import { Dispatcher, IGitHubUser } from '../../lib/dispatcher'
+import { Dispatcher } from '../../lib/dispatcher'
+import { IGitHubUser } from '../../lib/databases'
 import { IHistoryState } from '../../lib/app-state'
 import { ThrottledScheduler } from '../lib/throttled-scheduler'
