+++ app/src/ui/lfs/attribute-mismatch.tsx
@@ -4,7 +4,7 @@ import { ButtonGroup } from '../lib/button-group'
 import { Dialog, DialogContent, DialogFooter } from '../dialog'
 import { LinkButton } from '../lib/link-button'
 import { getGlobalConfigPath } from '../../lib/git'
-import { shell } from '../../lib/dispatcher/app-shell'
+import { shell } from '../../lib/app-shell'
 
 interface IAttributeMismatchProps {
   /** Called when the dialog should be dismissed. */
@@ -61,9 +61,11 @@ export class AttributeMismatch extends React.Component<
       <Dialog
         id="lfs-attribute-mismatch"
         title={
-          __DARWIN__
-            ? 'Update Existing Git LFS Filters?'
-            : 'Update existing Git LFS filters?'
+          __DARWIN__ ? (
+            'Update Existing Git LFS Filters?'
+          ) : (
+            'Update existing Git LFS filters?'
+          )
         }
         onDismissed={this.props.onDismissed}
         onSubmit={this.props.onUpdateExistingFilters}
@@ -79,9 +81,11 @@ export class AttributeMismatch extends React.Component<
         <DialogFooter>
           <ButtonGroup>
             <Button type="submit">
-              {__DARWIN__
-                ? 'Update Existing Filters'
-                : 'Update existing filters'}
+              {__DARWIN__ ? (
+                'Update Existing Filters'
+              ) : (
+                'Update existing filters'
+              )}
             </Button>
             <Button onClick={this.props.onDismissed}>
               {__DARWIN__ ? 'Not Now' : 'Not now'}
