+++ app/test/test-app-shell.ts
@@ -1,6 +1,6 @@
 /* tslint:disable:no-sync-functions */
 
-import { IAppShell } from '../src/lib/dispatcher/app-shell'
+import { IAppShell } from '../src/lib/app-shell'
 
 import * as Fs from 'fs'
