+++ app/test/unit/accounts-store-test.ts
@@ -2,7 +2,7 @@ import * as chai from 'chai'
 const expect = chai.expect
 
 import { Account } from '../../src/models/account'
-import { AccountsStore } from '../../src/lib/dispatcher'
+import { AccountsStore } from '../../src/lib/stores'
 import { InMemoryStore } from '../in-memory-store'
 import { AsyncInMemoryStore } from '../async-in-memory-store'
