+++ app/test/unit/git/branch-test.ts
@@ -10,7 +10,7 @@ import {
   IValidBranch,
   IUnbornRepository,
 } from '../../../src/models/tip'
-import { GitStore } from '../../../src/lib/dispatcher/git-store'
+import { GitStore } from '../../../src/lib/stores'
 import { shell } from '../../test-app-shell'
 import { GitProcess } from 'dugite'
